Rename OpynFAST to pyOpenFAST
